fresh implementation of react-router with centralized routes file. moved translation stuff from App.jsx to Home.jsx
