+++ package.json
@@ -22,7 +22,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-i18next": "^15.4.1",
-    "react-router-dom": "^7.3.0",
+    "react-router-dom": "^7.4.0",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
