+++ src/routes/index.jsx
@@ -3,6 +3,8 @@ import Home from "../pages/Home";
 import About from "../pages/About";
 import Contact from "../pages/Contact";
 import NotFound from "../pages/NotFound";
+import TestTranslations from "../pages/TestTranslations";
+
 const AppRoutes = () => {
   return (
     <Router>
@@ -10,6 +12,7 @@ const AppRoutes = () => {
         <Route path="/" element={<Home />} />
         <Route path="/about" element={<About />} />
         <Route path="/contact" element={<Contact />} />
+        <Route path="/test-translations" element={<TestTranslations />} />
         <Route path="*" element={<NotFound />} />
       </Routes>
     </Router>
